refactor(startup): extract startup model catalog helpers
Refs #933
