+++ crates/tau-coding-agent/src/model_catalog.rs
@@ -1,7 +1,6 @@
 #[cfg(test)]
 pub(crate) use tau_provider::default_model_catalog_cache_path;
 pub(crate) use tau_provider::{
-    ensure_model_supports_tools, load_model_catalog_with_cache, parse_models_list_args,
-    render_model_show, render_models_list, ModelCatalog, ModelCatalogLoadOptions,
-    MODELS_LIST_USAGE, MODEL_SHOW_USAGE,
+    parse_models_list_args, render_model_show, render_models_list, ModelCatalog, MODELS_LIST_USAGE,
+    MODEL_SHOW_USAGE,
 };
